refactor(browser): split AppBrowser, location as a pub field
